tower-city-hospital/src/main/java/org/wecancodeit/cleveland/hospital/Hospital.java
package org.wecancodeit.cleveland.hospital;

import java.util.Collection;
import java.util.HashMap;
import java.util.Map;

public class Hospital {
	Map<String, Employee> employees = new HashMap<>();
	// Map<key, value> plural

	void hire(Employee anyEmployee) {
		employees.put(anyEmployee.getNumber(), anyEmployee);
	}

	public Employee findEmployee(String number) {
		return employees.get(number);
	}

	public Collection<Employee> getAllEmployees() {
		return employees.values();
	}

	public void fire(Employee anyEmployee) {
		employees.remove(anyEmployee.getNumber(), anyEmployee);
	}

	public void showEmployees() {
		for (Employee employee : employees.values()) {
			System.out.println("Emp Number: " + employee.getNumber() + " Name" + employee.getName() + " Salary: "
					+ employee.calculatePay());
		}
	}

	public void showAllMedicalPersonnel() {
		for (Employee employee : employees.values()) {
			if(employee instanceof MedicalDuties) {
				System.out.println("Emp Number: " + employee.getNumber() + " Name" + employee.getName());
		}

	}
}
}